// src/app/shared/model/astroProfile.model.ts
export class AstroProfile {
    status: string;
    count: string;
    astroId: number;
    name: string;
    imgUrl: string;
    expert: string;
    expertise: { id: string, name: string }[];
    experience: string;
    gender: string
    lang: string;
    language: { id: string, name: string }[];
    description: string;
    followerCount: number;
    followerStatus: string;
    isFollow: boolean;
    remainingTime: string;
    talkedTime: string;
    avgRating: string;
    fName: string;
    lName: string;
    contactNo: string;
    email: string;
    isBusy: number;
    isActive: number;
    active: boolean;
    chatStatus: string;
    password: string;
    countryId: number;
    stateId: number;
    cityId: number;
    astroStatus: boolean;

}